requirements.txt
customtkinter==5.2.1
Pillow==10.2.0
pandas==2.2.0
openpyxl==3.1.2
fpdf==1.7.2
tkcalendar==1.6.1 